Fix salesman email to customer HTML test
The test was used to determine whether HTML tags can be inserted to an 
email sent to a customer using the current 1-to-1 (one customer, 
one PDF) email sending mechanism.

The test result was negative, since the Gmail client (used for the 
test) didn't interpret correctly the closing </strong> tag.

This test was mistakenly added and not deleted from the "Create a 
sign up form" commit.