Shorten PDF link for better usability
The PDF viewer can reference to a PDF file using an absolute or a 
relative path. Until this commit, an absolute path was used after 
the "Remove hard-coding to enable flexible deployment" commit.
Before that commit a relative path was used.

For example, before this commit, a PDF link looked like:
http://localhost:8080/sp/pdfjs/viewer.html?file=http://localhost:8080/sp/file/test-3pages-1test-3pagesyanivf-g
After this commit it will look like:
http://localhost:8080/sp/pdfjs/viewer.html?file=/sp/file/test-3pages-1test-3pagesyanivf-g