Replace old chat design with a new one
The new chat looks better, allows for better UI by implementing
the old chat features in a better way (e.g. whether a salesman is
online or offline), and enables new features such as a lean version
of "ask me a question".

The chat is based upon the following code snippet:
http://codepen.io/marcobiedermann/pen/rohwn