Add chat customization options & add powered by
- Enabled to set for each salesman related viewer whether the chat will
be enabled or not, and whether it will load opened or not
(for screens > 768 pixels).

By default the chat will be set to be enabled and closed.

This templating task actually fixed an undesired UX behavior where the
chat would automatically open or close based on the viewport
breakpoint, i.e. resizing the browser window automatically opened /
closed the chat.

- sp-chat.css was merged to sp-viewer.css for efficiency reasons.

- Added Powered by SlidePiper + link to the bottom left of the viewer.

DB operations:
ALTER TABLE `sales_men`
  ADD `viewer_is_chat_enabled` varchar(8) DEFAULT 'true',
  ADD `viewer_is_chat_open` varchar(8) DEFAULT NULL;