Replace customers import mechanism
The current mechanism lacks the ability to parse CSV values including 
a new line for example. Moreover, since the current mechanism doesn't 
even use a standard library for the task - it was replaced to the 
Apache Commons CSVParser which is a well accepted mechanism.

Moreover, each value is now trimmed for white spaces (and new lines) 
before it is inserted into the DB.

Lastly, each email is checked and must be a valid email address for 
the customer record to be entered into the DB.

DB operations (assuming company max length < 256):
ALTER TABLE picascrafxzhbcmd.customers CHANGE company company
VARCHAR(255);