Print a file link hash
The class goal is to help generate a default customer file link.

First introduced in version 0.4.0, a default customer file link is 
automatically created after a file has been uploaded.

Therefore, for a file uploaded prior to version 0.4.0, a manual 
process similar to DbLayer.setFileLinkHash() is needed to be conducted 
in order to create a default customer file link.

By manually inserting a file link record to the msg_info table and 
then setting the id field based on the result of this class we can 
create a default customer file link.