Create users CTA, fix pages viewed, refactor metrics
For some reason the pages viewed metric was removed from 
Analytics.sqlFilesData property - now it is has been restored.

Also changed the default metrics text to be N/A expect "Total views" 
which would be 0 - when there aren't any views.

Even if there are views, some metrics could still show N/A due to the 
metrics SQL queries, and the nature of when events are sent.