Create wizard to send documents to customers
Three-part wizard allowing the user to select their customers by email,
then the documents they would like to send. The final part of the wizard
renders the user choices, with the option to copy one document, or copy
all documents been sent to that user. This will allow the document to be
sent over any medium.

The second option is to send one document, or send all, to a particular 
customer. This functionality opens to default email window of the user,
with a pre-rendered mail body and subject.

Other changes: the request origin variable (requestOrigin) is passed 
through the routing emulator, so that when requests are made to the 
server to get customer () or files lists (), the request origin routes
this information to the right function - either to create the list of 
uploaded files or customers, or to the document wizard.

All AJAX requests are now split into the request, and all have a
callback function which handles the routing of the information.