Change dashboard scrollbar functionality
Remove usage of perfect scrollbar in favour of overflow-y property.

Add this CSS styling to the analytics menus, the documents and 
customers section, and to the send email wizard.

Add a max height of 300px to the nav bar container.

Add a min-height of 100px to the scrollbar thumb, and always display 
scrollbar.

Remove perfect scrollbar CDN, and remove all code relating to it.