Add search func to all customer and document lists
This change involved rewriting how the tables are created. Using
Datatables API, the tables are created with this API, instead of by 
writing HTML in the js code.

In send email wizard and documents & customers, the lists are
searchable. In order to open doc name in mailto, the sp-mailto class
was moved onto the file name, and assigning data-file-name attribute on
to it.

Remove all individual 'Copy' buttons on document wizard, have one
button to copy all file links. Change 'Copy All' button to 'Copy',
and change 'Send All' button to 'Send'.

Change text on navbar: 'Send Email' to 'Send Documents'.

Change text in Alert Email from 'Chat with Client' to 'Co-browse and
Chat with Client'.

Remove ability to click 'Upload Documents' if there is no document
selected; a toastr error will be shown.