Add sort on Marketing and Sales Analytics
Add ability to sort documents in Marketing Analytics by document name
and performance. Add ability to sort by customer name in Sales
Analytics. It is not possible yet for sorting by performance, but there
is a sweetalert saying that it is coming soon.

Fix issue where files were being requested when clicking on search or
sort in Marketing Analytics.

Add a timestamp on the customers table.

If no name exists for the customer, then display the customer's email
address in Sales Analytics.

Viewer - make youtube widget loop through the loaded video.

Viewer - create default SlidePiper viewer - add the settings to the SQL
query