Sort documents based on local timestamp
Create new db field (local_timestamp) which receives the local time and
date that a file was uploaded from the browser.

In DB, sort the files by descending timestamp, then by name.

Move 'Copy' and 'Send' buttons on Dashboard to the end of the row, and
align them in the centre.