Fix Calendly widget on mobile
With the previous implementation of the widget, it took the entire 
space of the mobile screen which caused an issue where it was 
impossible to close the widget.

With the current solution, not only that the widget can be closed via 
the Sweet Alert window, but it also paves the way towards a universal 
solution for implementing widgets in our viewer.