Add viewer customisation settings to signup page
Enable selection of different colours for the viewer toolbar and 
buttons. Also enable the ability to choose the text and links of the
CTA buttons, and to choose a logo for the viewer toolbar. 

Based on these selections, the other colours of the viewer toolbar
are decided in the java code.

If the user makes no choices, there is a default color and logo 
selection already in place for the viewer.