Fix bug stopping mailto from opening in safari
Safari does not allow window.open() methods unless they come from event
listeners. So have changed this method after signing up a new user to
location.href in order to open the mailto link.

Move the 'Contact Us' CTA to be CTA3 instead of CTA1.

Add default viewer_toolbar_logo_link if no logo in selected. If a logo
is selected, but no link is added, change the link to the link to view
that document.

Remove Company as a required field on the signup form.