Viewer - Make video and chat widget responsive
Make video widget responsive to page size, but also with a max and min
width.

Make video widget compatible with YouTube iFrame API, and with any
other type of video link.

Add a field viewer_widget1_is_youtube_video to check if the video is
from YouTube or not.

Make chat widget responsive to page size with a min and max width.

Add SP Terms, Conditions, and Privacy to right dropdown toolbar. This
disappears when screen is too small and video widget is fixed to the
left of the screen.