Viewer - Make changes to widget responsiveness
Add <hr> element back to video widget.

Set video max-width to 500px.

Remove top-left border of video widget when it is nested on the left
of the page.

Set max-width on toolbar logo, and CTA buttons, to 150px. 

Set new collapse width on CTA buttons, so as that at their max-width
they collapse before overlapping the toolbar logo.

Reduce height of calendar widget when no text and set min-height, and
set the icon size to 1.5em.