Make further fixes to calendly and video widget
Remove calendly widget animation and reduce size. It is now appearing
on all pages.

The fix to the calendly widget also fixed the video widget which did not
show on pages where it was not set to appear. This also affected
the chat. 

Fix bug on video widget so that it resizes properly on opening when
scrolling to the page on which it is set to open.