Revamp viewer
- Upgrade viewer to latest stable version (1.4.20 instead of 1.1.114)

- Pre-rendering of document pages upon load (currently set to 20 pages).
The defined number of pages (20) start to render when the document is 
loaded. Based on the device performance and the document richness of 
text and pictures it can take time until all of the 20 pages would 
render (just likes it takes a single page to render). Just to 
clarify, a rendered page is page that is completely shown after 
the spiral on a white page is shown, i.e. the rendering 
process)

- Add document scaling on desktop view

- Download correct file name

- Allow page thumbnails

- Fix overlapping pages

- Show loading bar

- Fix false online (green) status of chat

- Make sure that font settings can only be applied to CTA buttons, and 
not all of the toolbar text elements


Also:
- Hebrew (or any UTF-8 encoded string) can now be written in CTA 
buttons

- Start using StringEscapeUtils.escapeHtml4() for avoiding XSS attacks