Add animation and an extra line to Form Widget
Form widget has a similar, 'beating' animation to the Like Widget.

Add the option of a second line to the button text, which, when
appended to the right side button, is divided by a margin.

Add a 'pop' functionality, which allows the salesman to select after
how many seconds the form widget should open after the page loads.