Add splash page & refactor widget 3
Create splash page for the Viewer.

Add functionality to display the Question Widget at the bottom of the
document. It is possible to display the widget both on the right and
on the bottom.

Set the right side widget defaut color to be the CTA button color, but
the user can choose their own color in Dashboard settings.